feat: add constructs v10 compatibility (#150)
Closes https://github.com/cdk8s-team/cdk8s/issues/773

This is a draft PR that updates cdk8s to support constructs v3 and v10 as a peer dependencies, so it can work with both AWS CDK v1 and V2.

Out of the box, this does not mean all cdk8s apps and libraries can switch to using constructs v10, since in order to use constructs v10, all of the CDK libraries you are using must also support constructs v10. (So for example, you might also need to upgrade to a version of cdk8s-plus-22 that also supports constructs v10).