fix: Chart.of does not recognize valid charts (#402)
Fixes #401 

I haven't explicitly tested this with the scenario described in #401 but I believe the root cause is the same issue described [here](https://github.com/aws/constructs/pull/955). In a nutshell, "instanceof" is unreliable since it can cause issues if libraries are referencing different versions of cdk8s library in `node_modules`.

Signed-off-by: Christopher Rybicki <rybickic@amazon.com>