fix: align schedule class to K8s cron requirements (#705)
## Summary:

This class was initially created aligning with [Schedule Class]() which we utilize in CDK. Some of the features in that class does not makes sense in K8s. For instance, `rate()` would not work with K8s since it only expects a valid cron expression within its `schedule` field for `CronJob Resource`. 
Hence, this PR would align the Schedule Class with what is valid for K8s. 

## Changes Introduced:

* Removing `rate()` feature.
* Renaming class from `Schedule` to `CronSchedule` since K8s just supports cron expressions.
* Removing `year` parameter from `CronOptions` since it is not supported in K8s [CronJob Schedule field](https://kubernetes.io/docs/concepts/workloads/controllers/cron-jobs/#cron-schedule-syntax). 

## Commands Run:

`yarn && yarn compile && yarn test`

## Sign Off (DCO):

Signed-off-by: Vinayak Kukreja <vinakuk@amazon.com>