fix: app `outdir` property cannot be different from CLI or cdk8s.yaml file (#783)
Currently setting the `outdir` property for CDK8s app throws an error,
```
ERROR: synthesis failed, app expected to create "dist"
```

This is due to the CLI not able to recognize this change in the `outdir` in the application property. 
The value specified in the `outdir` needs to be similar to what is mentioned either in CLI's `--output` argument or the `output` property in the cdk8s.yaml file. 

Signed-off-by: Vinayak Kukreja <vinakuk@amazon.com>

Resolves https://github.com/cdk8s-team/cdk8s-cli/issues/210