feat(app): hook into synthesis with `chart.toJson` (#1053)
The `Chart` class provides a `toJson` method that can synthesize a single chart, and not the entire app. 

It stands to reason that customers will override this method if they need to inject custom logic to the synthesis process. However, the `toJson` method isn't actually being called by `app.synth`, and therefore doesn't provide a proper hook.

This PR standardizes the synth logic to use `chart.toJson` everywhere. 

### Misc

- Use `construct.node` instead of the deprecated `Node.of(construct)`