fix(metadata): shared collections are mutated unexpectedly  (#1223)
The `ApiObjectMetadataDefinition` class exposes a bunch of mutating methods on metadata collections (`labels`, `finalizer`, etc...). These methods mutate the instance variable, but the problem is that instance variable is initialized with a reference to a collection in the input properties. If this reference is shared between several instance (for example shared labels), a mutation on one instance affects other instances.

This PR performs a copy-on-write for those collection properties.

Fixes https://github.com/cdk8s-team/cdk8s/issues/791